C++ compiler cleanup: cast signed to unsigned

+++ Parser/tokenizer.c
@@ -1230,7 +1230,7 @@ tok_get(register struct tok_state *tok, char **p_start, char **p_end)
 		do {
 			*tp++ = c = tok_nextc(tok);
 		} while (c != EOF && c != '\n' &&
-			 tp - cbuf + 1 < sizeof(cbuf));
+			 (unsigned int)(tp - cbuf + 1) < sizeof(cbuf));
 		*tp = '\0';
 		for (cp = tabforms;
 		     cp < tabforms + sizeof(tabforms)/sizeof(tabforms[0]);
